backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
pdfplumber==0.11.0
PyMuPDF==1.23.26
python-docx==1.1.0
Pillow==10.3.0
pytesseract==0.3.10
spacy==3.7.4
openai==1.30.1
faiss-cpu==1.8.0
sentence-transformers==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
requests==2.31.0
google-generativeai==0.5.0
